requirements.txt
boto3==1.34.34
pytest==7.4.4
pytest-cov==4.1.0
moto==4.2.14
requests==2.31.0
PyJWT==2.8.0
coverage==7.4.1